+++ ttnn/cpp/pybind11/operations/__init__.hpp
@@ -18,6 +18,7 @@
 #include "ttnn/operations/pool/avgpool/avg_pool_pybind.hpp"
 #include "ttnn/operations/pool/maxpool/maxpool_pybind.hpp"
 #include "ttnn/operations/pool/maxpool/max_pool2d_pybind.hpp"
+#include "ttnn/operations/pool/upsample/upsample_pybind.hpp"
 #include "ttnn/operations/eltwise/binary/binary_pybind.hpp"
 #include "ttnn/operations/eltwise/ternary/ternary_pybind.hpp"
 #include "ttnn/operations/eltwise/binary_backward/binary_backward_pybind.hpp"
@@ -100,6 +101,7 @@ void py_module(py::module& module) {
     auto m_pool = module.def_submodule("pool", "pooling  operations");
     pool::py_module(m_pool);
     avgpool::py_module(m_pool);
+    upsample::py_module(m_pool);
 
     auto m_normalization = module.def_submodule("normalization", "normalization operations");
     normalization::py_module(m_normalization);
