+++ ttnn/cpp/ttnn/operations/data_movement.hpp
@@ -7,7 +7,7 @@
 #include "ttnn/tensor/types.hpp"
 #include "ttnn/cpp/ttnn/operations/data_movement/concat/device/concat_device_operation.hpp"
 #include "ttnn/deprecated/tt_dnn/op_library/repeat/repeat_op.hpp"
-#include "ttnn/operations/upsample/upsample_op.hpp"
+#include "ttnn/deprecated/tt_dnn/op_library/composite/composite_ops.hpp"
 #include "ttnn/operations/core/core.hpp"
 
 #include <ranges>
@@ -16,62 +16,6 @@ namespace ttnn {
 namespace operations {
 namespace data_movement {
 
-struct UpSample {
-    static ttnn::Tensor operator()(
-        const ttnn::Tensor& input_tensor,
-        std::variant<int, std::array<int, 2>, std::array<int, 3>, std::array<int, 4>> scale_factor,
-        std::optional<MemoryConfig> output_mem_config = std::nullopt) {
-        MemoryConfig mem_config = output_mem_config.value_or(ttnn::DRAM_MEMORY_CONFIG);
-
-        int scale_h = 1;
-        int scale_w = 1;
-        std::visit(
-            [&scale_h, &scale_w](auto&& sf) {
-                using T = std::decay_t<decltype(sf)>;
-                if constexpr (std::is_same_v<T, int>) {
-                    scale_h = sf;
-                    scale_w = sf;
-                } else if constexpr (std::is_same_v<T, std::array<int, 2>>) {
-                    scale_w = sf.at(0);
-                    int scale_c = sf.at(1);
-                    TT_FATAL(scale_c == 1);
-                } else if constexpr (std::is_same_v<T, std::array<int, 3>>) {
-                    scale_h = sf.at(0);
-                    scale_w = sf.at(1);
-                    int scale_c = sf.at(2);
-                    TT_FATAL(scale_c == 1);
-                } else if constexpr (std::is_same_v<T, std::array<int, 4>>) {
-                    int scale_n = sf.at(0);
-                    scale_h = sf.at(1);
-                    scale_w = sf.at(2);
-                    int scale_c = sf.at(3);
-                    TT_FATAL(scale_n == 1);
-                    TT_FATAL(scale_c == 1);
-                } else {
-                    // static_assert(false, "Unsupported scale factor");
-                    static_assert(sizeof(T) != 0, "Type check failed.");
-                }
-            },
-            scale_factor);
-
-        // DEBUG
-        // fmt::print("scale_h: {}, scale_w: {}\n", scale_h, scale_w);
-
-        if (input_tensor.is_sharded()) {
-            // TT_FATAL(not input_tensor.is_sharded());
-            int shard_height = input_tensor.memory_config().shard_spec.value().shape[0];
-            const auto batch_size = input_tensor.get_shape()[0];
-            const auto input_h = input_tensor.get_shape()[1];
-            const auto input_w = input_tensor.get_shape()[2];
-            const auto num_channels = input_tensor.get_shape()[3];
-            if (shard_height % input_w != 0) {
-                TT_FATAL(shard_height % input_w != 0);
-            }
-        }
-
-        return tt::tt_metal::upsample(input_tensor, scale_h, scale_w, mem_config);
-    }
-};
 
 struct Repeat {
     static ttnn::Tensor operator()(
@@ -88,7 +32,6 @@ struct Repeat {
 }  // namespace data_movement
 }  // namespace operations
 
-constexpr auto upsample = ttnn::register_operation_with_auto_launch_op<"ttnn::upsample", ttnn::operations::data_movement::UpSample>();
 constexpr auto repeat = ttnn::register_operation_with_auto_launch_op<"ttnn::repeat", ttnn::operations::data_movement::Repeat>();
 
 }  // namespace ttnn
