+++ ttnn/cpp/ttnn/operations/pool/upsample/device/kernels/dataflow/writer_upsample_unary_stick_layout_interleaved_start_id.cpp
