+++ ttnn/cpp/ttnn/operations/pool/upsample/device/upsample_op.cpp
