+++ ttnn/cpp/ttnn/operations/pool/upsample/upsample_pybind.hpp
@@ -1,0 +1,13 @@
+// SPDX-FileCopyrightText: © 2024 Tenstorrent Inc.
+//
+// SPDX-License-Identifier: Apache-2.0
+
+#pragma once
+
+#include "pybind11/pybind_fwd.hpp"
+
+namespace ttnn::operations::upsample {
+
+void py_module(py::module& module);
+
+} //namespace ttnn::operations::upsample
